+++ spec/support/mutations/mutation.rb
@@ -1,0 +1,7 @@
+require_relative './create_invoice_mutation'
+
+Mutation = GraphQL::ObjectType.define do
+  name 'Mutation'
+
+  field :createInvoice, field: CreateInvoiceMutation.field
+end